chore: improved help text and finish implementatio of user class
