+++ internal/app/model/user.go
@@ -6,6 +6,7 @@ import (
 	"io"
 	"log"
 	"net/http"
+	"sync"
 
 	"github.com/0bvim/goctobot/utils"
 )
@@ -18,12 +19,12 @@ const (
 // Main user struct
 type MyUser struct {
 	Followers  []User
-	Following  []User
-	Allowed    []User
 	Denied     []User
+	Following  []User
 	Login      string `json:"login"`
 	TargetUser string
 	Token      string
+	UserStatus map[string]string
 }
 
 // a single user struct
@@ -32,7 +33,13 @@ type User struct {
 }
 
 func (u *MyUser) FetchFollowing(count *int) {
-	url := fmt.Sprintf(FOLLOWING_URL, u.Login)
+	var url string
+	if u.TargetUser != "" {
+		url = fmt.Sprintf(FOLLOWING_URL, u.TargetUser)
+	} else {
+		url = fmt.Sprintf(FOLLOWING_URL, u.Login)
+	}
+
 	fetchData(url, "following", u, count)
 }
 
@@ -42,11 +49,31 @@ func (u *MyUser) FetchFollowers(count *int) {
 }
 
 func (u *MyUser) Unfollow() {
-	//TODO: Implement this function
+	var usersToUnfollow []string
+	for _, user := range u.Following {
+		if !userInList(user, u.Followers) || u.UserStatus[user.Login] == "Allow" {
+			usersToUnfollow = append(usersToUnfollow, user.Login)
+		}
+	}
+
+	processUsers(usersToUnfollow, "unfollow")
 }
 
 func (u *MyUser) Follow() {
-	//TODO: Implement this function
+	if u.TargetUser == "" {
+		fmt.Print("User to fetch? ")
+		fmt.Scanln(&u.TargetUser)
+	}
+	u.FetchFollowers(new(int))
+
+	var usersToFollow []string
+	for _, user := range u.Followers {
+		if !userInList(user, u.Following) || u.UserStatus[user.Login] == "Deny" {
+			usersToFollow = append(usersToFollow, u.Login)
+		}
+	}
+
+	processUsers(usersToFollow, "follow")
 }
 
 func fetchData(url, action string, u *MyUser, count *int) {
@@ -76,5 +103,91 @@ func fetchData(url, action string, u *MyUser, count *int) {
 		}
 
 		url = utils.GetNextURL(resp)
+	}
+}
+
+func processUsers(users []string, command string) {
+	var wg sync.WaitGroup
+	count := 0
+	for _, user := range users {
+		wg.Add(1)
+		if command == "unfollow" {
+			go unfollowUser(user, &count, &wg)
+		} else if command == "follow" {
+			go followUser(user, &count, &wg)
+		}
+		err := utils.LogFollowUnfollow(user, command)
+		if err != nil {
+			fmt.Println("Error loggin"+command+"action:", err)
+		}
+	}
+	wg.Wait()
+}
+
+func userInList(user User, list []User) bool {
+	for _, u := range list {
+		if u.Login == user.Login {
+			return true
+		}
+	}
+
+	return false
+}
+
+func unfollowUser(user string, count *int, wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	url := fmt.Sprintf("https://api.github.com/user/following/%s", user)
+	req, err := http.NewRequest("DELETE", url, nil)
+	if err != nil {
+		fmt.Printf("Error unfollowing user %s: %v\n", user, err)
+		return
+	}
+	req.Header.Set("Authorization", "token "+utils.GetToken())
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		fmt.Printf("Error unfollowing user %s: %v\n", user, err)
+		return
+	}
+	defer resp.Body.Close()
+
+	switch resp.StatusCode {
+	case http.StatusNoContent:
+		fmt.Printf("User: %s has been unfollowed!\n", user)
+	case http.StatusForbidden, http.StatusTooManyRequests:
+		utils.HandleRateLimit(count)
+		unfollowUser(user, count, wg)
+	default:
+		fmt.Printf("Error unfollowing %s: %v\n", user, resp.Status)
+	}
+}
+
+func followUser(user string, count *int, wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	url := fmt.Sprintf("https://api.github.com/user/following/%s", user)
+	req, err := http.NewRequest("PUT", url, nil)
+	if err != nil {
+		fmt.Printf("Error following user %s: %v\n", user, err)
+		return
+	}
+	req.Header.Set("Authorization", "token "+utils.GetToken())
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		fmt.Printf("Error following user %s: %v\n", user, err)
+		return
+	}
+	defer resp.Body.Close()
+
+	switch resp.StatusCode {
+	case http.StatusNoContent:
+		fmt.Printf("User: %s has been followed!\n", user)
+	case http.StatusForbidden, http.StatusTooManyRequests:
+		utils.HandleRateLimit(count)
+		followUser(user, count, wg)
+	default:
+		fmt.Printf("Error following %s: %v\n", user, resp.Status)
 	}
 }
